### 2021-03-10 #434 @danielvartan
mctq: An R Package for the Munich ChronoType Questionnaire
Submitting Author: Daniel Vartanian (@danielvartan)
Other Authors: Ana Amelia Benedito-Silva (@AnaAmeliaBeneditoSIlva) and Mario Pedrazzoli (@pedrazzo)
Repository: https://github.com/gipsousp/mctq
Version submitted: 0.0.0.9000

Version accepted: TBD

---

```
Package: mctq
Title: An R Package for the Munich ChronoType Questionnaire
Version: 0.0.0.9000
    c(person(given = "Daniel",
             family = "Vartanian",
             email = "danvartan@gmail.com",
             comment = c(ORCID = "0000-0001-7782-759X")),
      person(given = "Ana Amelia",
             family = "Benedito-Silva",
             email = "aamelia@usp.br",
             role = c("aut", "sad"),
      person(given = "Mario",
             family = "Pedrazzoli",
             email = "pedrazzo@usp.br",
             role = c("aut", "sad"),
             comment = c(ORCID = "0000-0002-5257-591X")),
      person(given = "Interdisciplinary Sleep Research Group (GIPSO)",
             role = c("fnd", "cph")),
      person(given = "University of Sao Paulo (USP)",
             role = c("fnd")))
Maintainer: Daniel Vartanian <danvartan@gmail.com>
Description: A complete and consistent toolkit to process the Munich ChronoType
    Questionnaire (MCTQ) for all of its three versions (standard, micro,
    and shift).
    checkmate,
    hms,
    lifecycle,
    lubridate
    covr,
    crayon,
    datasets,
    grDevices,
    mockr,
    utils
```

The Munich Chronotype Questionnaire (MCTQ) may look like a simple questionnaire, but it requires a lot of date/time manipulation. This can be really challenging, especially if you’re dealing with a large set of data. `mctq` provides reliable tools, thoroughly tested, to help scientists with that task.

`mctq` was designed for researchers in the field of sleep and chronobiology. The scientific applications of the package are the same as the scale it addresses, _i.e._ to assess peoples' sleep behavior and characteristics of circadian rhythms.

No other packages do the same thing (to my knowledge).

I don't think this is applicable to our package.

N/A
