+++ vignettes/mctq.Rmd
@@ -65,7 +65,7 @@ data <- readr::read_csv(mctq::raw_data("vignette_mctq.csv"),
 
 ## Converting your data
 
-`mctq` makes use of the [lubridate](https://lubridate.tidyverse.org/) and [hms](https://hms.tidyverse.org/) package, which provide special objects to deal with date/time values in R. If your dataset does not conform to this structure, you first need to convert your data to it.
+`mctq` makes use of the [lubridate](https://lubridate.tidyverse.org/) and [hms](https://hms.tidyverse.org/) packages, which provide special objects to deal with date/time values in R. If your dataset does not conform to this structure, you first need to convert your data to it.
 
 Due to the circular nature of time, we strongly recommend that you use appropriate temporal objects while dealing with date/time in R. That can help you get rid of several computation mistakes while trying to adapt your data from a base 10 to a system rooted in a base 12 numerical system.
 
