### 2021-03-20 #434 @jooolia
### Editor checks:

- [x] **Fit**: The package meets criteria for [fit](https://devguide.ropensci.org/policies.html#package-categories) and [overlap](https://devguide.ropensci.org/policies.html#overlap)covr
- [x] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

 #### Editor comments

 Thanks for the submission. I did some initial automated checks on the spelling, recommended practices, and the code coverage and everything looks shipshape! I will now proceed to looking for reviewers.

Thanks! Julia
