+++ man/mctq-package.Rd
@@ -9,8 +9,7 @@
 \if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
 
 A complete and consistent toolkit to process the Munich ChronoType 
-    Questionnaire (MCTQ) for all of its three versions (standard, micro, 
-    and shift).
+    Questionnaire (MCTQ) for its three versions (standard, micro, and shift).
 }
 \seealso{
 Useful links:
