+++ tests/testthat/test-raw_data.R
@@ -1,13 +1,7 @@
 test_that("raw_data() | general test", {
     checkmate::expect_character(raw_data(), any.missing = FALSE,
                                 all.missing = FALSE)
-    checkmate::expect_character(raw_data(raw_data()[1]), len = 1,
-                                any.missing = FALSE, all.missing = FALSE)
-
-    expect_equal(raw_data(NULL), dir(system.file("extdata", package = "mctq")))
-    expect_equal(raw_data(raw_data()[1]),
-                 system.file("extdata", raw_data()[1], package = "mctq",
-                             mustWork = TRUE))
+    checkmate::expect_string(raw_data(raw_data()[1]), na.ok = FALSE)
 })
 
 test_that("raw_data() | error test", {
