+++ tests/testthat/test-utils-mocks.R
@@ -7,6 +7,11 @@ test_that("is_namespace_loaded() | general test", {
     expect_equal(is_namespace_loaded(x), isNamespaceLoaded(x))
 })
 
+test_that("require_namespace() | general test", {
+    x <- "base"
+    expect_equal(require_namespace(x), requireNamespace(x, quietly = TRUE))
+})
+
 test_that("read_line() | general test", {
     expect_equal(read_line(""), "")
 })

[…]
