+++ _pkgdown.yml
@@ -66,6 +66,10 @@ articles:
 authors:
   Daniel Vartanian:
     href: https://github.com/danielvartan
+  Ana Amelia Benedito-Silva:
+    href: https://github.com/AnaAmeliaBeneditoSIlva
+  Mario Pedrazzoli:
+    href: https://github.com/pedrazzo
   Interdisciplinary Sleep Research Group (GIPSO):
     href: http://each.usp.br/gipso
   University of Sao Paulo (USP):
