+++ DESCRIPTION
@@ -24,7 +24,6 @@ Authors@R:
 Maintainer: Daniel Vartanian <danvartan@gmail.com>
 Description: A complete and consistent toolkit to process the Munich ChronoType 
     Questionnaire (MCTQ) for its three versions (standard, micro, and shift).
-    (CODEMETA WORKFLOW TEST).
 License: MIT + file LICENSE
 URL: https://gipsousp.github.io/mctq/, https://github.com/gipsousp/mctq/
 BugReports: https://github.com/gipsousp/mctq/issues/
