+++ tests/testthat/test-utils-mocks.R
@@ -2,11 +2,6 @@ test_that("is_interactive() | general test", {
     expect_equal(is_interactive(), interactive())
 })
 
-test_that("is_namespace_loaded() | general test", {
-    x <- "base"
-    expect_equal(is_namespace_loaded(x), isNamespaceLoaded(x))
-})
-
 test_that("require_namespace() | general test", {
     x <- "base"
     expect_equal(require_namespace(x), requireNamespace(x, quietly = TRUE))
