+++ .github/workflows/codemeta.yml
@@ -4,6 +4,7 @@ on:
       - master
     paths:
       - DESCRIPTION
+      - .github/workflows/codemeta.yml
 
 name: codemeta
 jobs:
@@ -14,8 +15,10 @@ jobs:
     steps:
       - uses: actions/checkout@v1
       - uses: r-lib/actions/setup-r@v1
-      - name: Install codemetar
-        run: Rscript -e 'install.packages("codemetar")'
+      - name: Install remotes
+        run: Rscript -e 'install.packages("remotes")'
+      - name: Install codemetar (dev version)
+        run: Rscript -e 'remotes::install_github("ropensci/codemetar")'
       - name: Render codemeta
         run: Rscript -e 'codemetar::write_codemeta()'
       - name: Commit results
