+++ tests/testthat.R
@@ -26,5 +26,10 @@ test_check("mctq")
 # goodpractice::gp()
 # spelling::spell_check_package()
 # spelling::update_wordlist()
+# usethis::use_tidy_description()
+# usethis::use_coverage()
+# usethis::use_pkgdown_github_pages()
+# usethis::use_github_action_check_full()
+# usethis::use_github_action_check_standard()
 # normalizePath(readClipboard(), "/", mustWork = FALSE)
 # usethis::use_logo("./.archive/ai/hex-logo.png")
