+++ .archive/vignettes/date-time.Rmd
@@ -1,21 +1,0 @@
----
-title: "Dealing with dates and time in R"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Dealing with dates and time in R}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-__UNDER DEVELOPMENT__
-
-Shows how time arithmetic can be tricky, possible ways to go around problems, and how the `mctq` package can help with this.
-
-Add "Visualizing date/time objects with ggplot2" topic.
