+++ .archive/vignettes/converting-data.Rmd
@@ -1,19 +1,0 @@
----
-title: "How to convert your MCTQ data"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{How to convert your MCTQ data}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-__UNDER DEVELOPMENT__
-
-Shows how to convert time data from decimals or radians to lubridate objects, as the inverse.
