+++ .archive/md/ropensci_submission.md
@@ -1,155 +1,0 @@
-Submitting Author: Daniel Vartanian (@danielvartan)
-Other Authors: Ana Amelia Benedito-Silva (@AnaAmeliaBeneditoSIlva) and Mario Pedrazzoli (@pedrazzo)
-Repository: https://github.com/gipsousp/mctq
-Version submitted: 0.0.0.9000
-Editor: <!--editor--> TBD <!--end-editor-->
-Reviewers: <!--reviewers-list--> TBD <!--end-reviewers-list-->
-<!--due-dates-list--><!--end-due-dates-list-->
-Archive: TBD
-Version accepted: TBD
-
----
-
-
-
--   Paste the full DESCRIPTION file inside a code block below:
-
-```
-Package: mctq
-Title: The R Package for the Munich ChronoType Questionnaire (MCTQ)
-Version: 0.0.0.9000
-Authors@R: 
-    c(person(given = "Daniel",
-             family = "Vartanian",
-             role = c("aut", "cre", "cph"),
-             email = "danvartan@gmail.com",
-             comment = c(ORCID = "0000-0001-7782-759X")),
-      person(given = "Ana Amelia",
-             family = "Benedito-Silva",
-             email = "aamelia@usp.br",
-             role = c("aut", "sad"),
-             comment = c(ORCID = "0000-0003-4976-2623")),
-      person(given = "Mario",
-             family = "Pedrazzoli",
-             email = "pedrazzo@usp.br",
-             role = c("aut", "sad"),
-             comment = c(ORCID = "0000-0002-5257-591X")),
-      person(given = "Interdisciplinary Sleep Research Group (GIPSO)",
-             role = c("fnd", "cph")),
-      person(given = "University of Sao Paulo (USP)",
-             role = c("fnd")))
-Maintainer: Daniel Vartanian <danvartan@gmail.com>
-Description: A complete and consistent toolkit to process Munich ChronoType 
-    Questionnaire (MCTQ) data in R for all of its three versions (standard, 
-    micro, and shift).
-License: MIT + file LICENSE
-URL: https://gipsousp.github.io/mctq/, https://github.com/gipsousp/mctq/
-BugReports: https://github.com/gipsousp/mctq/issues/
-Encoding: UTF-8
-LazyData: true
-Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.1.1
-Depends:
-    R (>= 3.6)
-Imports: 
-    checkmate,
-    dplyr,
-    hms,
-    lifecycle,
-    lubridate
-Suggests:
-    covr,
-    crayon,
-    datasets,
-    ggplot2,
-    grDevices,
-    knitr,
-    magrittr,
-    mockr,
-    readr,
-    rmarkdown,
-    spelling,
-    stats,
-    testthat (>= 2.0.0),
-    utils
-Config/testthat/edition: 2
-VignetteBuilder: knitr
-Language: en-US
-```
-
-
-## Scope
-
-- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below. If you are unsure, we suggest you make a pre-submission inquiry.):
-
-	- [ ] data retrieval
-	- [ ] data extraction
-	- [x] data munging
-	- [ ] data deposition
-	- [ ] workflow automation
-	- [ ] version control
-	- [ ] citation management and bibliometrics
-	- [ ] scientific software wrappers
-	- [ ] field and lab reproducibility tools
-	- [ ] database software bindings
-	- [ ] geospatial data
-	- [ ] text analysis
-
-- Explain how and why the package falls under these categories (briefly, 1-2 sentences):
-
[…]
