+++ .archive/vignettes/data-wrangling.Rmd
@@ -1,19 +1,0 @@
----
-title: "MCTQ data wrangling"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{MCTQ data wrangling}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-__UNDER DEVELOPMENT__
-
-Shows commons issues when dealing with raw MCTQ data, how to fix them, and how `mctq` can make it easy for researcher while dealing with this problems.
