### 2021-04-27 caa2e11
Fix typos in the documentation
