### 2021-04-30 #434 @danielvartan
Hi @jooolia,

Thanks for your effort.

There are few people I know in the fields of sleep and chronobiology who work with R. The first person that crossed my mind was @leocadio-miguel (Mario André Leocadio Miguel | ORCID: https://orcid.org/0000-0002-7248-3529). Mario and I have worked on some projects together. I don't know if this may be an issue for this kind of peer-review.

I will try to think of other possible candidates.
