+++ R/micro_mctq.R
@@ -212,7 +212,7 @@
 #'   R class: `Duration`.}
 #' }
 #'
-#' @source Created by Daniel Vartanian (package's author).
+#' @source Created by Daniel Vartanian (package author).
 #' @family datasets
 #' @template references_b
 "micro_mctq"
