+++ R/shift_mctq.R
@@ -466,7 +466,7 @@
 #'   R class: `Duration`.}
 #' }
 #'
-#' @source Created by Daniel Vartanian (package's author).
+#' @source Created by Daniel Vartanian (package author).
 #' @family datasets
 #' @template references_b
 "shift_mctq"
