+++ man/std_mctq.Rd
@@ -330,7 +330,7 @@ R class: \code{Duration}.}
 }
 }
 \source{
-Created by Daniel Vartanian (package's author).
+Created by Daniel Vartanian (package author).
 }
 \usage{
 std_mctq
