### 2021-05-22 #434 @jonkeane
Ok, I finished my review, which is below. I say this below, but it bears repeating: really great job on this package so far!

- [x] **Examples** (that run successfully locally) for all exported functions

---

This is a really nice package that has some nice functionality and abstractions for dealing with this kind of survey data. I enjoyed taking time to review it.

I'm very impressed with the amount of documentation, the authors have devoted a lot of time to making sure there is good documentation explaining what's going on which is fantastic!

#### General comments that apply to a number of places in the package

##### Object Orientation

There are a number of classes/methods uses, extends, or exports which is really great. I think that a number of areas in the package would benefit from leaning more into that / on that. There are a number of places where functions would use large `if/else` blocks effectively like method dispatch. I think the package would be easier to reason about, easier to maintain, and easier to extend if many of those were converted into methods with corresponding classes.

##### Variable naming convenctions

This is more stylistic, but there are frequently variables with names that don't have a concrete connection to what the value is that they store (e.g. `x`, `y`, `z`, `foo`). Some of these are inevitable, but I would recommend trying to replace these with more meaningful descriptions of what they are holding. It helps a lot with reading this code to have variables named with what they are rather than needing to remember what `x` is in this context.

##### The `convert()` and other functions that vary their output based on arguments

Making easy conversion functions and teaching people how to use them for times+dates is very important, and I laud the authors for the care they've put in to making functions that help with that. I do think, however, that the `convert()` function might be over loaded too much. It is itself pretty complicated, but it the output is highly dependent on the arguments (e.g. the `class` argument). I think it might be an easier-to-learn (and secondarily easier-to-maintain) to split this up into a bunch of smaller `as_class()` functions that are all documented together. IME it's easier to teach folks how to use functions like that since they do exactly what's on the label. This fits in with making [these functions pure](https://adv-r.hadley.nz/fp.html#functional-programming-languages) which isn't a requirement of course, but would be helpful in this situation. There are a few other functions like this which take arguments that vary their output classes that you might consider making pure, and is you have this robust set of `as*()` classes you could even recommend using them to vary the output as in: `as_difftime(shortest_interval(...))` instead of `shortest_interval(..., class = "difftime")`

##### Defensive programming

There is extensive use of defensive programming principles through out the package. I think this is really great, especially with a package that is targeted at an audience who might not be familiar with some of the more esoteric error messages (e.g. the classic "object of type closure is not subsetable") or might be more likely to supply incompatible objects as arguments. No change is needed, but I did want to call this out since it's clear a significant amount of time was put into making this happen!

##### Minor documentation changes

I've submitted a PR along with a number of minor documentation changes (I found that easier than listing out where each one was).

I've changed all instances of he/she to they — this is the preferred way to refer to unknown third persons in English, is less clunky, and doesn't reinforce the gender binary.

A few places you use "subject" and a few places "volunteer". I think either is fine (or you could use "respondent" since it's a survey, I find that to be the least loaded term for this), but it would be good to standardize on one.

You might consider only using th lifecycle badge at the package level instead of on each documentation page — it's a bit redundant unless some of your individual functions have a different lifecycle.

#### Specific comments about the code

Some of these are repeated/expanded explanations from the general points above. Many of them are minor points or stylistic changes, so hopefully aren't overwhelming.

##### `goodpractice::gp()` reports

```
fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically
    indicates Rd problems. LaTeX errors found: ! Please use \mathaccent for accents in math mode.
    \add@accent ... \let \ignorespaces \relax \accent #1 #2\egroup \spacefactor ... l.2116
    ...eqn{\textasciicircum{}{Shift}}{})}{napd} ! Missing { inserted. <to be read again> \egroup
    l.2116 ...eqn{\textasciicircum{}{Shift}}{})}{napd}
```

##### `sd()` masking
This is something that really should be avoided if possible. The name matching is nice, but that's such a core function that I worry this will cause a lot of confusion for users. Could you provide an escape hatch that dispatches `stats::sd` when the {mctq} arguments either aren't present or the unnamed arguments aren't {mctq} objects or rename it to something that won't clash? Alternatively, if you used S4 methods you could make `sd` a generic and then only dispatch when you have (the correct) {mctq} objects.

##### .archive folder
I found it a bit frustrating to search for things since I would sometimes get hits in the `.archive` folder. With git, all of that history is there now, would it be possible remove it going forward on the main branch?

##### Installation instructions
In the installation instructions in the readme and {mctq} vignette: you can use {remotes} instead of {devtools} if you want fewer dependencies at install time ({devtools} uses {remotes} under the hood.)

##### {mctq} vignette
in the {mctq} vignette, you use `library(magrittr)` though you also have a dependency on {dplyr} why not "just" import the pipe from {dplyr}?

##### On load message
On load message is nice, but maybe only needed for interactive sessions?

##### Storing values in variables when it's not needed/sometimes is confusing

I noticed a pattern of things like (one big example is in `test-sum_time.R`):

```
class <- "duration"
circular <- FALSE
vectorize <- FALSE
na.rm <- FALSE
object <- sum_time(t, u, v, w, x, y, z, class = class, circular = circular,
                   vectorize = vectorize, na.rm = na.rm)
expected <- lubridate::dhours(30)
expect_equal(object, expected) # 30:00:00
```
And then most of those variables get changed later (some do stay the same, of course). I find it much easier to follow tests like these if you do repeat the values and put them into the call:
```
object <- sum_time(t, u, v, w, x, y, z, class = "duration", circular = FALSE,
                   vectorize = FALSE, na.rm = FALSE)
expected <- lubridate::dhours(30)
expect_equal(object, expected) # 30:00:00
```
or even:
```
expect_equal(
  sum_time(t, u, v, w, x, y, z, class = "duration", circular = FALSE,
           vectorize = FALSE, na.rm = FALSE),
  lubridate::dhours(30)
) # 30:00:00
```
There's a bit more repetition of things like `FALSE` or `"duration"` but it's _much_ easier to read as someone trying to understand the code + it makes moving the tests around in this file (or to another file) muuuch easier because it's not reliant on what came before it.

##### The functionality of `midday_change()`

This might be my own out-of-area expertise, but `midday_change()` seems like an odd function. After reading the docs, I expected it would give me the day I'm on + 1 day and not just after the epoch. Is this a term-of-art (or transform-of-art?) that I'm not familiar with in chronobiology?

##### Assert the message (or a part of the message) on errors
Suggestion: There are a number of `expect_error(function(arg, arg))` I've found it helpful + caught a number of bugs myself when I assert error message itself (or some relevant portion of it)

##### test-utils.R
lines 246-247 look like they are duplicated from the test above (and not relevant to `shush()`)

##### Simplifying `swap()`
`swap()` could be simplified to `swap <- function(x, y) list(x = y, y = x)`, but if you don't want to I would recommend making `a` and `b` more explanatory (i.e. `first_arg` and `second_arg` respectively)

`swap_if()` you should avoid parse/evaling since that can lead to some very very strange errors it looks like the one use + the tests, this is always a relation between the two arguments that are given. Why not make the condition either be a logical (that's evaluated naturally) like `swap_if(x, y, condition = x > y)` or, you could make the condition a test, something like ``swap_if(x, y condition = `>`)`` which inside of it calls `condition(x, y)` so that you can change the relation between the two arguments.

##### `clock_roll()` method?
`clock_roll()` have you considered making this a method that dispatches based on the class of `x`?

##### `get_class()`
I don't think that `|| is.data.frame(x)` is ever called since data.frames inherit from lists (e.g. `> is.list(data.frame(x=1, y=2)); [1] TRUE`). `foo` could use a slightly more descriptive name, and is there any reason you don't use it in the `else` branch?

##### `str_extract_()`
is there any reason not to use `...` to pass most of the arguments to `regexpr`?

##### `dialog_line()`
looks like it has a lot of extra functionality — maybe this is needed for something in the future? Or was needed for something in the past, but it looks like it is just used in `qplot_walk` and then only to take control the flow of plots? And that dialog could be a bit clearer / use the more standard y/n (or the tidyverse style randomized 4 options where 3 are no and 1 is yes).

##### `sum_time()`
the vectorize argument name is a bit funny. I understand what it's doing when I read through the docs, but it wasn't immediately obvious to me what it was when I saw the argument name / in the tests. I wonder if it would be better to split this out into a different function (possibly called `vectorized_sum_time()` or `sum_times()`) rather than using an argument to change the behavior. Then one would know that `sum_time()` always returns one time, and the new function will return a vector of times, the length of all of it's inputs. That will also help adjust to the fact that they have different input requirements (e.g. that all `...` must have names in the vectorized version.)

##### Documentation simplificatoin
There are a number of functions that you have similar or the same notes sections for, have you considered using roxygen2's `@inheritSection` (or possibly templates, though I have less experience with those so I'm not 100% sure they would solve what you're looking for here) to write that in one place and place it in all of the others? It looks like you do use this in the docs for `sloss_week` it would be great to use this throughout

##### `__` in comments
This is stylistic: I don't think you need the `__` in the example comments (e.g. `#' ## __ To return `start` and `end` as list __`)

##### `assign_date()`
`assign_date()` the `return` argument + behavior is something that generally should be avoided — it means that this function can return objects of a few different shapes, and you'll have to parse through the arguments when reading code that uses it to know what those shapes are. I would consider splitting this into a few different functions (basically, one for each `return` type + whatever helpers you need to not repeat the processing code).

##### `convert()`
`convert()` has _a lot_ going on and can be overwhelming. I like that you have the helper/wrapper `conver_tu()` etc. but I would consider also breaking this function up. One thing that can be unexpected and trip people up is when a function returns different classes/shapes of objects depending on the input. One example of this is the `convert.data.frame()` method which returns a data.frame rather than a vector or scalar like all the others.

  This would likely be a pretty large overhaul, but I would consider leaning more into the various `as*()` methods that a lot of the convert codes uses (you could wrap some of them so that they are all of a similar style `as_logical`, `as_duration`, etc. and you could introduce your own timezone additions/assertions there too) would be easier to learn + use for people getting into this (and it makes the code written with them easier to follow because you know `as_duration()` will result in a vector of durations for example). You could still wrap all of those in `convert()` that then calls the appropriate `as*` method on the input and let method dispatch for the `as*` methods do the right thing for each input, which IMO would be easier to follow from a code-perspective as well. IME, it's actually easier to teach a bunch of single-use functions/methods like the propose `as*` group than it is to teach one function that does a lot based on its inputs.

  I also found the `orders` argument a little bit confusing, correct me if I'm wrong, but it's possible time values to parse (hours-minutes-seconds, hours-minutes, or hours), yeah? Maybe something like `accepted_values` might be a more obvious name for that? Or maybe `orders` is a term of art for this survey?

  Specifics:
  lines 269-271 this will only catch the logical-type `NA` but none of `NA_character_`, `NA_integer_`, etc. is that what you intend? Would `is.na(x)` be better here?

##### Startup message
For the startup message in zzz.R, should that be wrapped in `if(interactive()){}` or the like so that it only shows up during interactive sessions?

##### `fd()`
do you need both `assert_numeric_(wd)` and `checkmate::assert_numeric(wd, lower = 0, upper = 7)`, it seems like they are covering the same ground?

##### `gu()`
This is also true for all calls to this, `checkmate::assert_class("se", "hms")` results in the error: `Error: Assertion on '"se"' failed: Must inherit from class 'hms', but has class 'character'.` I know what that means, but do you think that would be something that the target users of {mctq} would know + be able to fix? You might consider making this more friendly, something like `"se" must be an hms object, and not a 'character'. See ?convert for help on converting objects`.

##### `chronotype()` definition
You should be able to simplify the definition of `chronotype()` to `chronotype <- msf_sc`

##### `pretty_mctq()`
In `pretty_mctq()` you (rightfully!) warn about rounding too early in the analysis process, I wonder if the default for the `round` argument here should be `FALSE` to encourage people to only round when they want to?

  FYI: It looks like [the approved solution for `where` not being exported from dplyr or tidyselect](https://github.com/r-lib/tidyselect/issues/201) is to use `utils::globalVariables("where")`

##### `ranomd_mctq.R`
The sample functions and environment assignment is pretty complicated to follow. I wonder if it would be easier to make your sampler functions accept the inputs you store already in the `values` lists and call them directly storing the variables in the scope — I suspect that would take ~the same amount of code and be much easier to follow (and won't have the same variable binding issues from the code analysis during R CMD check). If you want to keep the environments, did you consider using `parent.frame()` instead of passing the environment around?

##### `round_time()`
This is a bit less important than other examples, but this would be good as methods for each class — at least it would be easier to trace what the code path is for each class instead of needing to follow the if/else.

##### shortest versus longer interval
I'm surprised that `shortest_interval()` isn't paired with `longest_interval()` instead of what is currently `longer_interval()` is this a term of art for the survey? (or why it's not `shorter_interval()` and `longer_interval()`?)

This might be a larger pattern, but I wonder if instead of having a `class` argument that changes the output you encouraged people to do something like `as_hms(shortest_interval(...))` instead of `shortest_interval(..., class = "hms")`. This would make the `shortest_interval()` be able to return one kind of output and then the conversion is explicit to make for easier to read + interact with code.

##### `sum_time()`
Is it possible to make `sum` methods for these? That way someone doesn't need to know/use the specialized `sum_time(...)`, but can `sum(...)`

##### `check_length_one()`
This might not be the {checkmate} style, but you could use `assert_any_na()` in this instead of having to repeat the check + message for NAs at the beginning right? Any downsides to composing these check functions like that?

##### Mocking
utils-mocks.R — nice workaround for changes to mocking abilities!

##### `mdc()` simplification
`mdc <- function(x) midday_change(x)` could be `mdc <- midday_change`

#### Conclusion

Thanks for this really great package — I can see how this would be massively helpful for people running the MCTQ; especially those who are new(er) to R or have less experience handling datetime objects. The core functionality appears to all be there and is presented with lots of great documentation which is great.
