### 2021-05-22 #434 @danielvartan
Hi @jonkeane,

I'm very impressed with the extent and quality of your review. Thank you very much for such a prompt and thorough feedback. I'll start working on your suggestions today.

Thanks also for your PR. As you may have noticed, English is not my mother tongue. Although I made an effort to revise the text, some errors ended up passing.
