+++ _pkgdown.yml
@@ -70,6 +70,8 @@ authors:
     href: https://github.com/AnaAmeliaBeneditoSIlva
   Mario Pedrazzoli:
     href: https://github.com/pedrazzo
+  Jonathan Keane:
+    href: https://github.com/jonkeane
   Interdisciplinary Sleep Research Group (GIPSO):
     href: http://each.usp.br/gipso
   University of Sao Paulo (USP):
