+++ .archive/R/utils.R
@@ -1,261 +1,0 @@
-#' @family utility functions
-#' @noRd
-check_that_ <- function(data, ...) {
-
-    checkmate::assert_data_frame(data)
-
-    confront <- validate::check_that(data, ...)
-    validate::summary(confront)
-
-}
-
-#' Change dates by time of day
-#'
-#' @description
-#'
-#' `midday_change()` changes the dates of `POSIXt` objects accordingly to the
-#' time of day registered in the object values. The function do this by flatting
-#' the date to `1970-01-01` and them adding a day if the hour is lower than 12.
-#'
-#' `hms` objects are also allowed. In this case `midday_change()` will convert
-#' all values to `POSIXct` and do the same operations described above.
-#'
-#' This method can help with time charting, allowing a continuity after the
-#' midnight hour.
-#'
-#' `mdc()` is just a shorter version of `midday_change()` for convenience.
-#'
-#' @param x A `hms` or `POSIXt` object.
-#'
-#' @return A `POSIXct`object when `x` is `hms` or `POSIXct`, or a `POSIXlt`
-#'   object when `x` is `POSIXlt`.
-#'
-#' @family utility functions
-#' @noRd
-#'
-#' @examples
-#' x <- lubridate::ymd_hms("2021-01-15 20:02:01") # hour > 12h
-#' midday_change(x)
-#' #> [1] "1970-01-01 20:02:01 UTC" # Expected
-#'
-#' x <- lubridate::ymd_hms("1987-12-24 07:45:32") # hour < 12h
-#' midday_change(x)
-#' #> [1] "1970-01-02 07:45:32 UTC" # Expected
-midday_change = function(x) {
-    checkmate::assert_multi_class(x, c("hms", "POSIXct", "POSIXlt"))
-
-    x <- flat_posixt(convert(x, "posixct"))
-
-    x <- dplyr::case_when(
-        lubridate::hour(x) < 12 ~ change_day(x, 2),
-        TRUE ~ x
-    )
-
-    x
-}
-
-#' @rdname midday_change
-#' @noRd
-mdc <- function(x) midday_change(x)
-
-#' Flat dates of `POSIXt` objects
-#'
-#' @description
-#'
-#' `flat_posixt()` changes the dates of `POSIXt` objects to a base reference.
-#' This can be use to standardizing a point of origin to time values.
-#'
-#' Please note that, accordingly to `?base::POSIXct`, not all OSes can
-#' correctly convert times before 1902 or after 2037. It's best to maintain the
-#' base date as the [UNIX epoch](https://en.wikipedia.org/wiki/Unix_time),
-#' _i.e_ `"1970-01-01"` (default).
-#'
-#' @param x A `POSIXt` object.
-#' @param force_utc (optional) a `logical` value indicating if the time zone of
-#'   `x` must be forced to `"UTC"` (default: `TRUE`).
-#' @param base (optional) a string indicating the base date for flatting `x`
-#'   in `%Y-%m-%d` format (default: `"1970-01-01"`).
-#'
-#' @return An object of the same `POSIXt` class type as `x` with dates and time
-#'   zone transformed.
-#'
-#' @family utility functions
-#' @noRd
-#'
-#' @examples
-#' x <- lubridate::ymd_hms("2001-09-15 11:15:05", tz = "EST")
-#' flat_posixt(x)
-#' #> [1] "1970-01-01 11:15:05 UTC" # Expected
-#' flat_posixt(x, force_utc = FALSE)
-#' #> [1] "1970-01-01 11:15:05 EST" # Expected
-#' flat_posixt(x, base = "2020-01-01")
-#' #> [1] "2020-01-01 11:15:05 UTC"
[…]
