+++ .archive/R/validate_mctq.R
@@ -1,365 +1,0 @@
-#' Validate MCTQ values
-#'
-#' @description
-#'
-#' `r lifecycle::badge("experimental")`
-#'
-#' __UNDER DEVELOPMENT__
-#'
-#' @details
-#'
-#' __UNDER DEVELOPMENT__
-#'
-#' @section Custom Variables:
-#'
-#' __UNDER DEVELOPMENT__
-#'
-#' @param data A data frame.
-#' @param check (optional) a character vector indicating the variables to check.
-#'   See Details section for possible values. If `NULL` the function will
-#'   validate all MCTQ variables it can find on data (default: `NULL`).
-#' @param flag (optional) a logical value indicating if the function must return
-#'   only a logical value indicating if critical errors was found (default:
-#'   `FALSE`).
-#' @param custom (optional) a named list with characters values indicating the
-#'   equivalent names for the MCTQ variables. See Custom variables section to
-#'   learn more (default: `NULL`).
-#' @param breaks (optional) a logical value indicating if the function must stop
-#'   and ask to continue between each validation. This option only works in
-#'   interactive mode and when `flag = FALSE` (default: `TRUE`).
-#'
-#' @return If `flag` is `FALSE`, print results from validation tests. If `flag =
-#'   TRUE`, returns a logical value indicating if critical errors was found.
-#'
-#'   When run in interactive mode and with `remove_breaks = FALSE`,
-#'   `validate_mctq` offers a way to export the results.
-#'
-#' @family utility functions
-#' @importFrom validate in_range is_complete all_complete
-#' @importFrom lubridate is.duration
-#' @noRd
-#'
-#' @references
-#'
-#' Van der Loo, M., & De Jonge, E. (2018).
-#' _Statistical data cleaning with applications in R_. Hooboken, NJ: John
-#' Wiley & Sons. \doi{10.1002/9781118897126}.
-#'
-#' @examples
-#' \dontrun{
-#' validate_mctq(mctq::std_mctq)
-#' }
-validate_mctq <- function(data, check = NULL, flag = FALSE, custom = NULL,
-                          breaks = TRUE) {
-    # To do --------------------
-    #
-    # * Give the user a option to choose types of validation
-    # * Use ellipsis to assign types for validations
-
-    # Check arguments -----
-
-    checkmate::assert_data_frame(data, all.missing = FALSE, min.rows = 1,
-                                 min.cols = 1)
-    checkmate::assert_character(check, any.missing = FALSE, min.len = 1,
-                                unique = TRUE, null.ok = TRUE)
-    checkmate::assert_flag(flag)
-    checkmate::assert_list(custom, c("character", "NULL"), min.len = 1,
-                           unique = TRUE, names = "unique", null.ok = TRUE)
-    checkmate::assert_flag(breaks)
-
-    choices <- c(names(mctq::std_mctq))
-    checkmate::assert_subset(check, choices, empty.ok = TRUE)
-    checkmate::assert_subset(names(custom), choices, empty.ok = TRUE,
-                             .var.name = "custom argument list names")
-    checkmate::assert_subset(unlist(custom, use.names = FALSE),
-                             names(data), empty.ok = TRUE,
-                             .var.name = "custom argument list elements")
-
-    if (is.null(check) && is.null(custom) && !any(choices %in% names(data))) {
-        rlang::abort("`data` do not have any MCTQ variable to validate")
-        if(isTRUE(flag)) return(FALSE)
-    } else if (!is.null(check) && is.null(custom) &&
-               !any(check %in% names(data))) {
[…]
