+++ .archive/R/NEWS.md
@@ -1,11 +1,0 @@
-# mctq (development version)
-
-> Nothing to report for now.
-
-# mctq 1.0.0
-
-First stable `mctq` version (go go `mctq`!).
-
-# mctq 0.0.0.9000
-
-* Added a `NEWS.md` file to track changes to the package.

[…]
