+++ R/zzz.R
@@ -1,4 +1,7 @@
 .onAttach <- function(...) {
     tip <- "Learn how to use 'mctq' at gipsousp.github.io/mctq ."
-    packageStartupMessage(paste(strwrap(tip), collapse = "\n"))
+
+    if (is_interactive()) {
+        packageStartupMessage(paste(strwrap(tip), collapse = "\n"))
+    }
 }
