### 2021-05-23 ee974cf
Change installation instructions on README.md/.Rmd and mctq.Rmd vignette
