### 2021-05-23 fbaf698
Remove lines not relevant to `shush()` in test-utils.R
