### 2021-05-23 c404b07
Optimize `get_class()` function
