### 2021-05-23 5e38ef9
Add the `magrittr` pipe from `dplyr`to NAMESPACE
