+++ NAMESPACE
@@ -42,6 +42,7 @@ export(sloss_week)
 export(so)
 export(sum_time)
 export(tbt)
+importFrom(dplyr,"%>%")
 importFrom(dplyr,across)
 importFrom(hms,hms)
 importFrom(lifecycle,deprecate_soft)
