### 2021-05-23 a3343e9
Remove `magrittr` dependence from data-raw scripts
