+++ data-raw/micro_mctq.R
@@ -5,7 +5,6 @@
 # library(dplyr)
 # library(hms)
 # library(lubridate)
-# library(magrittr)
 # library(mctq)
 # library(rlang)
 # library(usethis)
