### 2021-05-23 17a6d9e
Remove `__` from the examples sections
