+++ R/fd.R
@@ -33,7 +33,8 @@
 #' @export
 #'
 #' @examples
-#' ## __ Scalar example __
+#' ## Scalar example
+#'
 #' fd(5)
 #' #> [1] 2 # Expected
 #' fd(4)
@@ -41,7 +42,8 @@
 #' fd(as.numeric(NA))
 #' #> [1] NA # Expected
 #'
-#' ## __ Vector example __
+#' ## Vector example
+#'
 #' fd(0:7)
 #' #> [1] 7 6 5 4 3 2 1 0 # Expected
 #' fd(c(1, NA))
