+++ R/tbt.R
@@ -68,7 +68,8 @@
 #' @export
 #'
 #' @examples
-#' ## __ Scalar example __
+#' ## Scalar example
+#'
 #' bt <- hms::parse_hm("22:10")
 #' gu <- hms::parse_hm("06:15")
 #' tbt(bt, gu)
@@ -84,7 +85,8 @@
 #' tbt(bt, gu)
 #' #> [1] NA # Expected
 #'
-#' ## __ Vector example __
+#' ## Vector example
+#'
 #' bt <- c(hms::parse_hm("23:50"), hms::parse_hm("02:30"))
 #' gu <- c(hms::parse_hm("09:30"), hms::parse_hm("11:25"))
 #' tbt(bt, gu)
