### 2021-05-23 d54c3bf
Remove `mdc()` function
