### 2021-05-23 dccc4c5
Change the `shortest_interval()` function name to `shorter_interval()`
