+++ R/shorter_interval.R
@@ -1,15 +1,15 @@
-#' Find the shortest interval between two hours
+#' Find the shorter interval between two hours
 #'
 #' @description
 #'
 #' `r lifecycle::badge("maturing")`
 #'
-#' `shortest_interval()` finds and returns the shortest interval between two
+#' `shorter_interval()` finds and returns the shorter interval between two
 #' `hms` or `POSIXt` object hours.
 #'
-#' `longer_interval()` do the inverse of `shortest_interval()`, _i.e_,
+#' `longer_interval()` do the inverse of `shorter_interval()`, _i.e_,
 #' finds the longer interval between two hours. It's just a wrapper for
-#' `shortest_interval(x, y, class, inverse = TRUE)`.
+#' `shorter_interval(x, y, class, inverse = TRUE)`.
 #'
 #' @details
 #'
@@ -19,21 +19,21 @@
 #' there will be always two possible intervals between them, as illustrated
 #' below.
 #'
-#' To figure out what interval is the  shortest or the longer,
-#' `shortest_interval()` verify two scenarios: 1. When `x` comes before `y`; and
+#' To figure out what interval is the  shorter or the longer,
+#' `shorter_interval()` verify two scenarios: 1. When `x` comes before `y`; and
 #' 2. when `x` comes after `y`. This only works if `x` value is smaller than
 #' `y`, therefore, the function will make sure to swap `x` and `y` values if the
 #' latter assumption is not true.
 #'
-#' Because `shortest_interval()` objective is to find the shortest interval, if
-#' `x` and `y` are equal, the shortest interval will have a length of 0 hours,
+#' Because `shorter_interval()` objective is to find the shorter interval, if
+#' `x` and `y` are equal, the shorter interval will have a length of 0 hours,
 #' resulting in an interval from `x` to `x`. But, if `inverse = TRUE` or
 #' `longer_interval()` is used instead, the latter condition will return a
 #' interval with 24 hours of length (from `x` to `x` + 1 day).
 #'
 #' In cases when `x` and `y` distance themselves by 12 hours, there will be no
-#' shortest or longer interval (they will have equal length). In those cases,
-#' `shortest_interval()` and `longer_interval()` will return the same value
+#' shorter or longer interval (they will have equal length). In those cases,
+#' `shorter_interval()` and `longer_interval()` will return the same value
 #' (an interval of 12 hours).
 #'
 #' ```
@@ -42,14 +42,14 @@
 #'    06:00              22:00     06:00              22:00
 #' -----|------------------|---------|------------------|----->
 #'               16h           8h             16h
-#'           longer int.  shortest int.   longer int.
+#'           longer int.  shorter int.   longer int.
 #'
 #'               day 1                      day 2
 #'      y                   x       y                   x
 #'    13:00               08:00   13:00               08:00
 #' -----|-------------------|-------|-------------------|----->
 #'               19h           5h            19h
-#'           longer int.  shortest int.  longer int.
+#'           longer int.  shorter int.  longer int.
 #'
 #'     x,y             x,y             x,y             x,y
 #'      x               y               x               y
@@ -75,7 +75,7 @@
 #'
 #' ## `class` argument
 #'
[…]
