+++ man/random_mctq.Rd
@@ -83,7 +83,7 @@ Other utility functions:
 \code{\link{qplot_walk}()},
 \code{\link{raw_data}()},
 \code{\link{round_time}()},
-\code{\link{shortest_interval}()},
+\code{\link{shorter_interval}()},
 \code{\link{sum_time}()}
 }
 \concept{utility functions}
