+++ man/round_time.Rd
@@ -77,7 +77,7 @@ Other utility functions:
 \code{\link{qplot_walk}()},
 \code{\link{random_mctq}()},
 \code{\link{raw_data}()},
-\code{\link{shortest_interval}()},
+\code{\link{shorter_interval}()},
 \code{\link{sum_time}()}
 }
 \concept{utility functions}
