+++ man/raw_data.Rd
@@ -40,7 +40,7 @@ Other utility functions:
 \code{\link{qplot_walk}()},
 \code{\link{random_mctq}()},
 \code{\link{round_time}()},
-\code{\link{shortest_interval}()},
+\code{\link{shorter_interval}()},
 \code{\link{sum_time}()}
 }
 \concept{utility functions}
