+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   mctq: mctq.html
   missing-sections: missing-sections.html
   sjl: sjl.html
-last_built: 2021-05-23T03:13Z
+last_built: 2021-05-23T04:34Z
 urls:
   reference: https://gipsousp.github.io/mctq//reference
   article: https://gipsousp.github.io/mctq//articles
