+++ R/zzz.R
@@ -1,7 +1,4 @@
 .onAttach <- function(...) {
     tip <- "Learn how to use 'mctq' at gipsousp.github.io/mctq ."
-
-    if (is_interactive()) {
-        packageStartupMessage(paste(strwrap(tip), collapse = "\n"))
-    }
+    package_startup_message(paste(strwrap(tip), collapse = "\n"))
 }
