+++ R/mctq-package.R
@@ -6,7 +6,7 @@
 ## usethis namespace: start
 #' @importFrom dplyr across %>%
 #' @importFrom lifecycle deprecate_soft
-#' @importFrom lubridate duration period %within%
+#' @importFrom lubridate duration period interval %within%
 #' @importFrom hms hms
 ## usethis namespace: end
 NULL
