+++ NAMESPACE
@@ -48,4 +48,5 @@ importFrom(hms,hms)
 importFrom(lifecycle,deprecate_soft)
 importFrom(lubridate,"%within%")
 importFrom(lubridate,duration)
+importFrom(lubridate,interval)
 importFrom(lubridate,period)
