+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   mctq: mctq.html
   missing-sections: missing-sections.html
   sjl: sjl.html
-last_built: 2021-05-23T04:41Z
+last_built: 2021-05-23T22:56Z
 urls:
   reference: https://gipsousp.github.io/mctq//reference
   article: https://gipsousp.github.io/mctq//articles
