### 2021-05-24 e0c8b0a
Change version number and require lubridate (>= 1.7.9)
