### 2021-05-24 5f7d970
Fix `convert_to_seconds()` difftime transformation
