### 2021-05-24 dd91589
Transform `round_time()` to a S3 generic.
