+++ NAMESPACE
@@ -10,6 +10,9 @@ S3method(convert,data.frame)
 S3method(convert,difftime)
 S3method(convert,hms)
 S3method(convert,numeric)
+S3method(round_time,default)
+S3method(round_time,difftime)
+S3method(round_time,hms)
 export(assign_date)
 export(chronotype)
 export(convert)
