### 2021-05-24 4ca993d
Update `round_time()` documentation
