### 2021-05-24 e81fc49
Fix `sjl.R` examples sections
