### 2021-05-24 46c056c
Add changes to the NEWS.md file
