+++ NEWS.md
@@ -1,3 +1,7 @@
+# mctq 0.0.0.9001
+
+* The DESCRIPTION file now requires a lubridate version >= 1.7.9. `vctrs` support was added in this version, fixing some issues related to rounding `Duration`, `Period`, and `Interval` objects. Click [here](https://github.com/tidyverse/lubridate/pull/871) to learn more.
+
 # mctq 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
