+++ R/convert.R
@@ -78,7 +78,7 @@
 #' object, `character` and `numeric` inputs cannot have time values equal or
 #' greater than 24 hours.
 #'
-#' That limits the set of `convert()` applications (_e.g_, when you want to
+#' That limits the set of `convert()` applications (e.g., when you want to
 #' parse a `character` to a `Duration` object of 35 hours). To get around this,
 #' some exceptions were made to `orders` __equal__ to `"H"`, `"M"`, `"S"`,
 #' `"HM"`, or `"HMS"`. For `"HM"` and `"HMS"` exceptions, minutes and seconds
@@ -92,7 +92,7 @@
 #'
 #' Operations with data frames are only column-wise and can be made by selecting
 #' specific columns (using the `col` argument) or a group of columns (by
-#' applying a flag function (_e.g_ `is.numeric`) in the `where` argument).
+#' applying a flag function (e.g., `is.numeric`) in the `where` argument).
 #'
 #' ## Different outputs
 #'
@@ -112,11 +112,11 @@
 #'
 #' `convert()` will return a [base::as.numeric()] output if `class` is set to
 #' `"numeric"` or `"double"`. For `Date` objects, the output will be the total
-#' of days since '1970-01-01' (UNIX epoch date). For date-time objects (_e.g._,
+#' of days since '1970-01-01' (UNIX epoch date). For date-time objects (e.g.,
 #' `POSIXt`), the output will be the total of seconds from the UNIX epoch
 #' (`1970-01-01 00:00:00 UTC`) (See
 #' [Unix time](https://en.wikipedia.org/wiki/Unix_time) to learn more). For time
-#' objects, (_e.g._, `hms`) the output will be the total of seconds.
+#' objects, (e.g., `hms`) the output will be the total of seconds.
 #'
 #' The output `class = "numeric"` can also be different if `input_unit` and
 #' `output_unit` are assigned.
@@ -160,7 +160,7 @@
 #'   days or 31557600 seconds).
 #' @param close_round (optional) a `logical` value indicating if numbers with
 #' decimals starting with three leading 0s or 9s must be rounded
-#' (_e.g._, 1.999) (default: `TRUE`).
+#' (e.g., 1.999) (default: `TRUE`).
 #' @param cols (optional) a `character` object indicating the column names in
 #'   `x` to transform (default: `NULL`).
 #' @param where (optional) a function to apply in a [tidyselect::where()] call
