+++ R/sum_time.R
@@ -47,7 +47,7 @@
 #'
 #' ## Timeline irregularities
 #'
-#' This function does not take into account timeline irregularities (_e.g._,
+#' This function does not take into account timeline irregularities (e.g.,
 #' leap years, DST, leap seconds). This may not be an issue for most people, but
 #' it must be considered when doing time arithmetic.
 #'
