### 2021-05-25 75a2d17
Change `qplot_walk()` documentation
