+++ man/qplot_walk.Rd
@@ -91,12 +91,24 @@ converts \code{Duration}, \code{Period}, and \code{difftime} objects to \code{hm
 \subsection{Midday change}{
 
 Time variables with values greater than \code{22:00:00} will automatically be
-converted to \code{POSIXct} with a midday change, i.e., all values with more
-than 12 hours will represent day one, and all the rest will represent day 2.
-This is made to better represent values that cross the midnight
-hour.
+converted to \code{POSIXct} and be attached to a two-day timeline using the midday
+hour as a cutting point, i.e., all values with 12 hours or more will be
+placed on day 1, and all the rest will be placed on day 2.
 
-You can disable this behavior using \code{midday_change = FALSE}.
+This is made to better represent time vectors that cross the midnight hour.
+You can disable this behavior by using \code{midday_change = FALSE}.
+
+Example: Say you have a vector of time values that cross the midnight hour
+(e.g., an \code{hms} vector with \code{22:00}, \code{23:00}, \code{00:00}, \code{01:00} values). If
+you use \code{midday_change = FALSE}, your data will be represented linearly.\preformatted{00:00 01:00                                22:00 23:00
+  |-----|------------------------------------|-----|------->
+}
+
+By using \code{midday_change = TRUE} (default), \code{qplot_walk()} will fit your data
+to a circular time frame of 24 hours.\preformatted{             day 1                         day 2
+                22:00 23:00 00:00 01:00
+------------------|-----|-----|-----|---------------------->
+}
 }
 
 \subsection{\code{id} variables}{
