### 2021-05-25 ffce945
Transform `na_as()` to an internal S3 generic
