+++ NAMESPACE
@@ -10,6 +10,17 @@ S3method(convert,data.frame)
 S3method(convert,difftime)
 S3method(convert,hms)
 S3method(convert,numeric)
+S3method(na_as,Date)
+S3method(na_as,Duration)
+S3method(na_as,POSIXct)
+S3method(na_as,POSIXlt)
+S3method(na_as,Period)
+S3method(na_as,character)
+S3method(na_as,difftime)
+S3method(na_as,hms)
+S3method(na_as,integer)
+S3method(na_as,logical)
+S3method(na_as,numeric)
 S3method(round_time,default)
 S3method(round_time,difftime)
 S3method(round_time,hms)
