### 2021-05-25 3c563ef
Add comments related to `check_numeric_()` and `assert_numeric_()`
