### 2021-05-25 ae4f3a9
Transform `clock_roll()` to an internal S3 generic
