+++ NAMESPACE
@@ -1,5 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(clock_roll,Duration)
+S3method(clock_roll,Period)
+S3method(clock_roll,difftime)
+S3method(clock_roll,hms)
 S3method(convert,Date)
 S3method(convert,Duration)
 S3method(convert,Interval)
