### 2021-05-25 ac55804
Optimize and join `swap()` and `swap_if()`
