### 2021-05-31 f0f9aaa
Update test suite
