+++ tests/testthat/test-random_mctq.R
@@ -1,6 +1,3 @@
-# Don't forget to run devtools::load_all(".") and uncomment the variables
-# before trying to run the tests interactively.
-
 test_that("random_mctq() | general test", {
     set.seed(1)
     checkmate::expect_list(shush(random_mctq(model = "standard")))
@@ -16,10 +13,12 @@ test_that("random_mctq() | general test", {
 })
 
 test_that("random_mctq() | error test", {
+    # ## Don't forget to run devtools::load_all(".") and uncomment the variables
+    # ## before trying to run the tests interactively.
+    #
     # require_namespace <- mctq:::require_namespace
     # random_mctq <- mctq::random_mctq
 
-    # "This function requires the `stats` package to run [...]"
     mock <- function(.parent = parent.frame(), .env = topenv(.parent)) {
         mockr::with_mock(
             require_namespace = function(...) FALSE,
@@ -27,22 +26,16 @@ test_that("random_mctq() | error test", {
     }
 
     # mock()
-    expect_error(mock())
+    expect_error(mock(), "This function requires the 'stats' package to run. ")
 
-    # Invalid values for `model` and `quiet`
-    expect_error(random_mctq(model = 1))
-    expect_error(random_mctq(quiet = 1))
+    expect_error(random_mctq(model = 1), "Assertion on 'model' failed")
+    expect_error(random_mctq(quiet = 1), "Assertion on 'quiet' failed")
 })
 
 test_that("random_mctq() | message test", {
-    # "\nModel: Standard MCTQ\n"
-    expect_message(random_mctq(model = "standard"))
-
-    # "\nModel: Micro MCTQ\n"
-    expect_message(random_mctq(model = "micro"))
-
-    "\nModel: MCTQ Shift\n"
-    expect_message(random_mctq(model = "shift"))
+    expect_message(random_mctq(model = "standard"), "\nModel: Standard MCTQ\n")
+    expect_message(random_mctq(model = "micro"), "\nModel: Micro MCTQ\n")
+    expect_message(random_mctq(model = "shift"), "\nModel: MCTQ Shift\n")
 })
 
 test_that("random_std_mctq() | general test", {
@@ -96,169 +89,258 @@ test_that("random_shift_mctq() | general test", {
     expect_false(x$nap_f_n)
 
     # Invalid values for `n_w` and `n_f`
-    expect_error(random_shift_mctq(n_w = ""))
-    expect_error(random_shift_mctq(n_w = c(1, 1)))
-    expect_error(random_shift_mctq(n_f = ""))
-    expect_error(random_shift_mctq(n_f = c(1, 1)))
+    expect_error(random_shift_mctq(n_w = ""), "Assertion on 'n_w' failed")
+    expect_error(random_shift_mctq(n_w = c(1, 1)), "Assertion on 'n_w' failed")
+    expect_error(random_shift_mctq(n_f = ""), "Assertion on 'n_f' failed")
+    expect_error(random_shift_mctq(n_f = c(1, 1)), "Assertion on 'n_f' failed")
 })
 
 test_that("normalize() | general test", {
     # else if (check_2)
-    min <- hms::parse_hm("22:00") + lubridate::ddays(1)
-    max <- hms::parse_hm("05:00")
-    mean <- hms::parse_hm("02:00")
-    object <- normalize(min, max, mean)
-    expected <- hms::hms(as.numeric(hms::parse_hm("02:00") +
-                                        lubridate::ddays()))
-    expect_equal(object$mean, expected)
+    object <- normalize(hms::parse_hm("22:00") + lubridate::ddays(1),
+                        hms::parse_hm("05:00"),
+                        hms::parse_hm("02:00"))
+    expect_equal(object$mean,
[…]
