### 2021-05-31 3925068
Add `require_pkg()` function
