+++ tests/testthat/test-utils-dialogs.R
@@ -5,7 +5,6 @@ test_that("dialog_line() | general test", {
     # is_interactive <- mctq:::is_interactive
     # require_namespace <- mctq:::require_namespace
     # read_line <- mctq:::read_line
-    # dialog_line <- mctq:::dialog_line
 
     mock <- function(.parent = parent.frame(), .env = topenv(.parent)) {
         mockr::with_mock(
@@ -13,8 +12,8 @@ test_that("dialog_line() | general test", {
             dialog_line(1))
     }
 
-    # x <- mock()
-    expect_equal(mock(), 999)
+    # mock()
+    expect_null(mock())
 
     mock <- function(.parent = parent.frame(), .env = topenv(.parent)) {
         mockr::with_mock(
@@ -22,24 +21,25 @@ test_that("dialog_line() | general test", {
             dialog_line(1, abort = TRUE))
     }
 
-    # x <- mock()
-    expect_equal(mock(), 999)
+    # mock()
+    expect_null(mock())
 
     mock <- function(.parent = parent.frame(), .env = topenv(.parent)) {
         mockr::with_mock(
             is_interactive = function(...) TRUE,
             require_namespace = function(...) TRUE,
             read_line = function(...) TRUE,
-            dialog_line(1, space_above = TRUE, space_below = TRUE))
+            dialog_line(1, combined_styles = "red", space_above = TRUE,
+                        space_below = TRUE))
     }
 
-    # x <- mock()
+    # mock()
     expect_equal(utils::capture.output(mock()), c("", "", "[1] TRUE"))
 })
 
 test_that("dialog_line() | error test", {
     expect_error(dialog_line(), "Assertion on 'list\\(...\\)' failed")
-    expect_error(dialog_line(1, combined_styles = ""),
+    expect_error(dialog_line(1, combined_styles = 1),
                  "Assertion on 'combined_styles' failed")
     expect_error(dialog_line(1, space_above = ""),
                  "Assertion on 'space_above' failed")
@@ -56,6 +56,13 @@ test_that("alert() | general test", {
     expect_message(alert(c(1, 2)), "12")
     expect_message(alert(1, 2), "12")
 
+    expect_warning(alert(1, type = "warning"), "1")
+
+    expect_equal(getElement(evaluate_promise(
+        alert(1, combined_styles = NULL, type = "cat")),
+        "output"),
+        "1")
+
     # ## Don't forget to run devtools::load_all(".") and uncomment the variables
     # ## before trying to run the tests interactively.
     #
@@ -67,7 +74,7 @@ test_that("alert() | general test", {
             alert("test"))
     }
 
-    # x <- mock()
+    # mock()
     expect_message(mock())
 })
 
@@ -75,6 +82,23 @@ test_that("alert() | error test", {
     expect_error(alert(), "Assertion on 'list\\(...\\)' failed")
     expect_error(alert(1, combined_styles = ""),
                  "Assertion on 'combined_styles' failed")
+    expect_error(alert(1, type = ""),
+                 "Assertion on 'type' failed")
     expect_error(alert(1, abort = ""),
+                 "Assertion on 'abort' failed")
+})
+
+test_that("crayonize() | general test", {
+    expect_equal(crayonize(1, abort = TRUE), NULL)
+    expect_equal(crayonize("test", combined_styles = NULL), "test")
+    expect_equal(crayonize("test", combined_styles = "red"),
+                 crayon::red("test"))
+})
+
+test_that("crayonize() | error test", {
+    expect_error(crayonize(), "Assertion on 'list\\(...\\)' failed")
+    expect_error(crayonize(1, combined_styles = ""),
+                 "Assertion on 'combined_styles' failed")
+    expect_error(crayonize(1, abort = ""),
                  "Assertion on 'abort' failed")
 })
