### 2021-05-31 166acfe
Optimize `require_pkg()` function
