### 2021-06-02 2abbb9f
Add the '>= 0.2' requirement to the dplyr import (to import the pipe operator)
