### 2021-06-02 818ae23
Update `raw_data()`
