+++ R/raw_data.R
@@ -7,31 +7,34 @@
 #' `mctq` comes bundled with raw fictional datasets for testing and learning
 #' purposes. `raw_data()` makes it easy to access their paths.
 #'
-#' @param file A string indicating the file name of the raw dataset. If `NULL`,
-#'   all raw dataset file names will be printed (default: `NULL`).
+#' @param file (optional) a `character` object indicating the raw data file
+#'   name(s). If `NULL`, all raw data file names will be printed (default:
+#'   `NULL`).
 #'
-#' @return If `path = NULL`, returns a `character` object with all raw dataset
-#'   file names available. Else, it returns the `file` path.
+#' @return
+#'
+#' * If `file = NULL`, a `character` object with all file names available.
+#' * If `file != NULL`, a string with the file name path.
 #'
 #' @family utility functions
 #' @export
 #'
 #' @examples
 #' \dontrun{
-#' ## To list all raw dataset file names available
+#' ## To list all raw data file names available
 #'
 #' raw_data()
 #'
-#' ## To get the file path from a specific raw dataset
+#' ## To get the file path from a specific raw data
 #'
 #' raw_data(raw_data()[1])
 #'
 #' raw_data("std_mctq.csv")}
 raw_data <- function(file = NULL) {
-    checkmate::assert_string(file, null.ok = TRUE)
+    checkmate::assert_character(file, min.len = 1, null.ok = TRUE)
 
     if (is.null(file)) {
-        dir(system.file("extdata", package = "mctq"))
+        list.files(system.file("extdata", package = "mctq"))
     } else {
         system.file("extdata", file, package = "mctq", mustWork = TRUE)
     }
