### 2021-06-03 137fc1c
Substitute `remotes` for `pak` in README
