+++ tests/testthat.R
@@ -7,26 +7,26 @@ test_check("mctq")
 # # For development use only (comment the code after use (Ctrl + Shift + C)).
 #
 # .rs.restartR()
-# devtools::install()
-# devtools::document()
-# devtools:test()
-# devtools::check()
-# covr::package_coverage()
-# covr::codecov(token = "")
-# pkgdown::build_site(preview = TRUE)
-# pkgdown::build_favicons(overwrite = TRUE)
-# pkgdown::build_article("mctq")
-# pkgdown::build_reference(preview = FALSE)
-# codemetar::write_codemeta()
 # codemetar::give_opinions()
-# urlchecker::url_check()
+# codemetar::write_codemeta()
+# covr::package_coverage()
+# devtools::check()
+# devtools::document()
+# devtools::install()
+# devtools:test()
 # goodpractice::gp()
+# normalizePath(readClipboard(), "/", mustWork = FALSE)
+# pkgdown::build_article("mctq")
+# pkgdown::build_favicons(overwrite = TRUE)
+# pkgdown::build_reference(preview = FALSE)
+# pkgdown::build_site(preview = TRUE)
 # spelling::spell_check_package()
 # spelling::update_wordlist()
-# usethis::use_tidy_description()
+# urlchecker::url_check()
 # usethis::use_coverage()
-# usethis::use_pkgdown_github_pages()
-# usethis::use_github_action_check_full()
-# usethis::use_github_action_check_standard()
+# usethis::use_github_action("check-pak")
+# usethis::use_github_action("pkgdown-pak")
+# usethis::use_github_action("test-coverage-pak")
 # usethis::use_logo("./.archive/ai/hex-logo.png")
-# normalizePath(readClipboard(), "/", mustWork = FALSE)
+# usethis::use_pkgdown_github_pages()
+# usethis::use_tidy_description()
