+++ tests/testthat/test-utils.R
@@ -95,24 +95,18 @@ test_that("change_day() | general test", {
 })
 
 test_that("single_quote_() | general test", {
-    expect_equal(single_quote_("test"), paste0("'", "test", "'"))
+    expect_equal(single_quote_("a"), paste0("'", "a", "'"))
     expect_equal(single_quote_(1), paste0("'", 1, "'"))
-    expect_equal(single_quote_(lubridate::dhours()),
-                 paste0("'", lubridate::dhours(), "'"))
 })
 
 test_that("double_quote_() | general test", {
-    expect_equal(double_quote_("test"), paste0("\"", "test", "\""))
+    expect_equal(double_quote_("a"), paste0("\"", "a", "\""))
     expect_equal(double_quote_(1), paste0("\"", 1, "\""))
-    expect_equal(double_quote_(lubridate::dhours()),
-                 paste0("\"", lubridate::dhours(), "\""))
 })
 
 test_that("backtick_() | general test", {
-    expect_equal(backtick_("test"), paste0("`", "test", "`"))
+    expect_equal(backtick_("a"), paste0("`", "a", "`"))
     expect_equal(backtick_(1), paste0("`", 1, "`"))
-    expect_equal(backtick_(lubridate::dhours()),
-                 paste0("`", lubridate::dhours(), "`"))
 })
 
 test_that("class_collapse() | general test", {
@@ -156,7 +150,7 @@ test_that("inline_collapse() | error test", {
 })
 
 test_that("shush() | general test", {
-    expect_equal(shush("test", quiet = FALSE), "test")
+    expect_equal(shush("a", quiet = FALSE), "a")
 
     test <- function() {
         warning("test", call. = FALSE)
