+++ .github/FUNDING.yml
@@ -1,1 +1,0 @@
-github: gipsousp
