+++ README.Rmd
@@ -20,6 +20,7 @@ library(hms)
 # mctq <a href='https://gipsousp.github.io/mctq'><img src='man/figures/logo.png' align="right" height="139" /></a>
 
 <!-- badges: start -->
+[![Buy Me A Coffee donate button](https://img.shields.io/badge/buy%20me%20a%20coffee-donate-yellow.svg)](https://www.buymeacoffee.com/danielvartan)
 [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/434_status.svg)](https://github.com/ropensci/software-review/issues/434)
 [![CRAN status](https://www.r-pkg.org/badges/version/mctq)](https://CRAN.R-project.org/package=mctq)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
@@ -150,3 +151,13 @@ citation("mctq")
 Please note that `mctq` is released with a [Contributor Code of
 Conduct](https://gipsousp.github.io/mctq/CODE_OF_CONDUCT.html). By contributing
 to this project, you agree to abide by its terms.
+
+## Donate
+
+[!["Buy Me A Coffee"](https://www.buymeacoffee.com/assets/img/custom_images/orange_img.png)](https://www.buymeacoffee.com/danielvartan)
+
+Working with science in Brazil is a daily challenge. There are few funding opportunities available and their value is not enough to live on. Added to this, every day Brazilian science suffers from deep cuts in funding, which requires researchers to always look for other sources of income.
+
+If this package helps you in any way or you simply want to support the author work, please consider donating or even creating a membership subscription (if you can!). Your support will help with the author scientific pursuit and with the package maintenance.
+
+To make a donation, click on the buymeacoffee.com badge above.
