### 2021-06-14 #434 @leocadio-miguel
Dear Editors and Authors, I have finished my review. Congratulations on this great work!
Please find below my comments.
Thank you for this opportunity!

- [x] **Examples** (that run successfully locally) for all exported functions

---

It was a pleasure to evaluate the MCTQ R package, authored by Vartanian, Benedito-Silva and Pedrazzoli. At first, It sounds interesting to use a specific R package to perform what seems to be a simple task - to process the results from a single questionnaire. However, almost nothing is simple when it comes to chronobiology, particularly because our research field deals with variables in the time domain, both using clock time and durations. This package is particularly useful to perform computations using time/date variables, to be able to read/write and derivate variables in the time domain and, also importantly, to make sure that all the proper conversion of time/date is performed.
As an expert in chronobiology and a final user of R (far from being a developer or basically understand the structure of R packages), I dedicated my review process to try to understand the package, to follow the step-by-step tutorial (get started introduction to mctq) using the sample data and, finally, to test the package using my own data extracted from my MCTQ database.
After carefully studying the package I was surprised with all the careful documentation, not only enough for the chronobiologist user but also complete, easy to follow and far from boring. As an example of care, the ?std_mctq link on the first section of the tutorial gives the reader full access to the not so intuitive variable descriptions, with a complete description of each variable including name, type of variable and R class. Moreover, all the tutorial steps ran smoothly and all the computed/derived variables were created and included in the dataset, as well as all proper conversions were performed. One of the most interesting features in the package is the convert() function. It adds a "second layer" of computations, allowing the final user to perform the conversions using a human-friendly syntax.
Finally, I was able to test the mctq package using my own data, converting time units properly and generating all the variables needed to report the results from the MCTQ.
I would like to congratulate the Authors and to acknowledge the importance of their initiative to make our (chronobiologists) lives a lot easier! Also, I would like to congratulate rOpenSci initiative to make the review process clear, transparent and accessible for the broad audience.
