### 2021-06-14 0536998
Add Mario Leocadio-Miguel as reviewer and remove GIPSO as a copyright holder
