+++ LICENSE.md
@@ -1,6 +1,6 @@
 # MIT License
 
-Copyright (c) 2021 Daniel Vartanian and Interdisciplinary Sleep Research Group (GIPSO)
+Copyright (c) 2021 Daniel Vartanian
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
