+++ _pkgdown.yml
@@ -72,6 +72,8 @@ authors:
     href: https://github.com/pedrazzo
   Jonathan Keane:
     href: https://github.com/jonkeane
+  Mario Andre Leocadio-Miguel:
+    href: https://github.com/leocadio-miguel
   Interdisciplinary Sleep Research Group (GIPSO):
     href: http://each.usp.br/gipso
   University of Sao Paulo (USP):
