### 2021-06-15 #434 @danielvartan
Thank you very much for your feedback @leocadio-miguel!

I already added both @leocadio-miguel and @jonkeane as reviewers to the package documentation (changes implemented in gipsousp/mctq@0536998 and gipsousp/mctq@5fa5233) (see also https://gipsousp.github.io/mctq/authors.html). Thanks again for yours excellent reviews!

I'm still working on the suggestions. You can track my progress by looking at the commit history. I believe that I can finish this work by the end of the month.
