### 2021-06-16 54270d5
Fix typos in the 'mctq' vignette
