### 2021-07-06 #434 @danielvartan
Hi @jooolia, @jonkeane, and @leocadio-miguel.

Some major tasks recently knocked on my door and I couldn't take the time needed to address all the reviewers' suggestions. I would like to know if this submission can be put on [hold](https://devguide.ropensci.org/policies.html?q=hold#policiesreviewprocess) for now.

I already worked on most of the suggestions, but I would like to take some time to review all the changes before I submit them here. At the moment my schedule is pretty much full, I believe I will have some time to finish this in the first week of August when I will have some spare time.

I'm sorry for the delay. I will try my best to submit all answers here ASAP.
