+++ R/sloss_week.R
@@ -18,7 +18,7 @@
 #' * If you are visualizing this documentation in plain text (`ASCII`), you may
 #' have some trouble understanding the equations. If you want a better viewer,
 #' you can see this documentation on the package
-#' [website](https://gipsousp.github.io/mctq/reference/).
+#' [website](https://gipso.github.io/mctq/reference/).
 #'
 #' ## Computation
 #'
