+++ R/zzz.R
@@ -1,4 +1,4 @@
 .onAttach <- function(...) {
-    tip <- "Learn how to use 'mctq' at gipsousp.github.io/mctq ."
+    tip <- "Learn how to use {mctq} at gipso.github.io/mctq ."
     package_startup_message(paste(strwrap(tip), collapse = "\n"))
 }
