+++ tests/testthat.R
@@ -20,6 +20,7 @@ test_check("mctq")
 # pkgdown::build_favicons(overwrite = TRUE)
 # pkgdown::build_reference(preview = FALSE)
 # pkgdown::build_site(preview = TRUE)
+# pkgnet::CreatePackageReport('mctq')
 # spelling::spell_check_package()
 # spelling::update_wordlist()
 # urlchecker::url_check()
