+++ R/micro_mctq.R
@@ -40,7 +40,7 @@
 #' addition to be a suggested data structure for MCTQ data.
 #'
 #' You can see the `micro_mctq` build and data wrangling processes
-#' [here](https://github.com/gipsousp/mctq/blob/master/data-raw/micro_mctq.R).
+#' [here](https://github.com/gipso/mctq/blob/master/data-raw/micro_mctq.R).
 #'
 #' ## Variable naming
 #'
