### 2021-10-11 #434 @danielvartan
I would like to apologize for the delay in responding to reviewers' comments. I had some urgent tasks to handle and I wanted to take some time to be able to devote myself to the suggestions pointed out before submitting my answer.

I also would like to thank the editor @jooolia and both referees @jonkeane and @leocadio-miguel for dedicating their time to reviewing our package. I have tried my best to incorporate all suggestions raised. Below I provide responses with the corresponding commit links.

---

## In response to suggestions made by @jonkeane

> Ok, I finished my review, which is below. I say this below, but it bears repeating: really great job on this package so far!

__Response__: Thank you very much @jonkeane. I have tried my best to incorporate almost all of your suggestions.

> ## Package Review
>
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
> * **Briefly describe any working relationship you have (had) with the package authors.**
>
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
>
> The package includes all the following forms of documentation:
>
> * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
> * [x]  **Function Documentation:** for all exported functions
> * [x]  **Examples** (that run successfully locally) for all exported functions
> * [x]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
> #### Functionality
>
> * [x]  **Installation:** Installation succeeds as documented.
> * [x]  **Functionality:** Any functional claims of the software been confirmed.
> * [ ]  **Performance:** Any performance claims of the software been confirmed.
> * [x]  **Automated tests:** Unit tests cover essential functions of the package
>   and a reasonable range of inputs and conditions. All tests pass on the local machine.
> * [x]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
>
> Estimated hours spent reviewing: 8.5
>
> * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

__Response__: I was pleased to add you as a reviewer. Changes implemented in gipso/mctq@5fa5233.

> ### Review Comments
>
> This is a really nice package that has some nice functionality and abstractions for dealing with this kind of survey data. I enjoyed taking time to review it.
>
> I'm very impressed with the amount of documentation, the authors have devoted a lot of time to making sure there is good documentation explaining what's going on which is fantastic!
>
> #### General comments that apply to a number of places in the package
>
> ##### Object Orientation
>
> There are a number of classes/methods uses, extends, or exports which is really great. I think that a number of areas in the package would benefit from leaning more into that / on that. There are a number of places where functions would use large `if/else` blocks effectively like method dispatch. I think the package would be easier to reason about, easier to maintain, and easier to extend if many of those were converted into methods with corresponding classes.

__Response__: Agreed. That's the result of some bad practices I'm still trying to quit. I already made the changes you suggested.

> ##### Variable naming convenctions
>
> This is more stylistic, but there are frequently variables with names that don't have a concrete connection to what the value is that they store (e.g. `x`, `y`, `z`, `foo`). Some of these are inevitable, but I would recommend trying to replace these with more meaningful descriptions of what they are holding. It helps a lot with reading this code to have variables named with what they are rather than needing to remember what `x` is in this context.

__Response__: I tried to change some of the variable names, but, as you said, some of these are inevitable.

> ##### The `convert()` and other functions that vary their output based on arguments
>
> Making easy conversion functions and teaching people how to use them for times+dates is very important, and I laud the authors for the care they've put in to making functions that help with that. I do think, however, that the `convert()` function might be over loaded too much. It is itself pretty complicated, but it the output is highly dependent on the arguments (e.g. the `class` argument). I think it might be an easier-to-learn (and secondarily easier-to-maintain) to split this up into a bunch of smaller `as_class()` functions that are all documented together. IME it's easier to teach folks how to use functions like that since they do exactly what's on the label. This fits in with making [these functions pure](https://adv-r.hadley.nz/fp.html#functional-programming-languages) which isn't a requirement of course, but would be helpful in this situation. There are a few other functions like this which take arguments that vary their output classes that you might consider making pure, and is you have this robust set of `as*()` classes you could even recommend using them to vary the output as in: `as_difftime(shortest_interval(...))` instead of `shortest_interval(..., class = "difftime")`

__Response__: Agreed. The `class` argument was a mistake. I removed it from all functions. Now all `mctq` functions are [pure](https://adv-r.hadley.nz/fp.html#functional-programming-languages) and [type-/size-stable](https://vctrs.r-lib.org/articles/stability.html#definitions), with only a few exceptions (like `qplot_walk()`). I comment more about the `convert()` function in a topic you signalized below.

> ##### Defensive programming
>
> There is extensive use of defensive programming principles through out the package. I think this is really great, especially with a package that is targeted at an audience who might not be familiar with some of the more esoteric error messages (e.g. the classic "object of type closure is not subsetable") or might be more likely to supply incompatible objects as arguments. No change is needed, but I did want to call this out since it's clear a significant amount of time was put into making this happen!

__Response__: Thanks for noticing! 🙂

> ##### Minor documentation changes
>
> I've submitted a PR along with a number of minor documentation changes (I found that easier than listing out where each one was).
>
> I've changed all instances of he/she to they — this is the preferred way to refer to unknown third persons in English, is less clunky, and doesn't reinforce the gender binary.

__Response__: I already approved the PR. Thanks again! Changes implemented in gipso/mctq@f4f9f57.

> A few places you use "subject" and a few places "volunteer". I think either is fine (or you could use "respondent" since it's a survey, I find that to be the least loaded term for this), but it would be good to standardize on one.

__Response__: Agreed. Changes implemented in gipso/mctq@d99b734.

> You might consider only using th lifecycle badge at the package level instead of on each documentation page — it's a bit redundant unless some of your individual functions have a different lifecycle.

__Response__: I find that is better to be a little redundant in this case. Some utility functions may need some maturing before I consider them as `stable`. Not the core functions thought. I plan to assign them as `stable` after the review process ends.

> #### Specific comments about the code
>
> Some of these are repeated/expanded explanations from the general points above. Many of them are minor points or stylistic changes, so hopefully aren't overwhelming.
>
> ##### `goodpractice::gp()` reports
>
> ```
> fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically
>     indicates Rd problems. LaTeX errors found: ! Please use \mathaccent for accents in math mode.
>     \add@accent ... \let \ignorespaces \relax \accent #1 #2\egroup \spacefactor ... l.2116
>     ...eqn{\textasciicircum{}{Shift}}{})}{napd} ! Missing { inserted. <to be read again> \egroup
>     l.2116 ...eqn{\textasciicircum{}{Shift}}{})}{napd}
> ```

__Response__: I wasn't able to reproduce this warning while running `goodpractice::gp()` or R CMD Check (the CI checks did not show the warning either). My output when running `goodpractice::gp()`: `♥ Ha! Wicked package! Keep up the rad work!`. I think this may be an issue with your `roxygen2` or `LaTeX` installation. Please check if this is the case.

> ##### `sd()` masking
>
> This is something that really should be avoided if possible. The name matching is nice, but that's such a core function that I worry this will cause a lot of confusion for users. Could you provide an escape hatch that dispatches `stats::sd` when the {mctq} arguments either aren't present or the unnamed arguments aren't {mctq} objects or rename it to something that won't clash? Alternatively, if you used S4 methods you could make `sd` a generic and then only dispatch when you have (the correct) {mctq} objects.

__Response__: Tough call, but you're right. That's a too dangerous masking. I decided to change the name to avoid the name collision you mentioned. This function is now named `sdu`. I also renamed the `ms()` function, as it was colliding with an important function of the `lubridate` package. Changes implemented in gipso/mctq@ec34d83 and gipso/mctq@1de1e2c.

> ##### .archive folder
>
> I found it a bit frustrating to search for things since I would sometimes get hits in the `.archive` folder. With git, all of that history is there now, would it be possible remove it going forward on the main branch?

__Response__: Sorry for that. I already deleted the `.archive` folder. Changes implemented in gipso/mctq@b047297.

> ##### Installation instructions
>
> In the installation instructions in the readme and {mctq} vignette: you can use {remotes} instead of {devtools} if you want fewer dependencies at install time ({devtools} uses {remotes} under the hood.)

__Response__: Agreed. Changes implemented in gipso/mctq@ee974cf.

> ##### {mctq} vignette
>
> in the {mctq} vignette, you use `library(magrittr)` though you also have a dependency on {dplyr} why not "just" import the pipe from {dplyr}?

__Response__: You're right (🤦️). Changes implemented in gipso/mctq@5e38ef9 and gipso/mctq@a3343e9.

> ##### On load message
>
> On load message is nice, but maybe only needed for interactive sessions?

__Response__: Agreed. In the end, I decided to remove the startup message (too much pollution). Changes implemented in gipso/mctq@72fd8b5.

> ##### Storing values in variables when it's not needed/sometimes is confusing
>
> I noticed a pattern of things like (one big example is in `test-sum_time.R`):
>
> ```
> class <- "duration"
> circular <- FALSE
> vectorize <- FALSE
> na.rm <- FALSE
> object <- sum_time(t, u, v, w, x, y, z, class = class, circular = circular,
>                    vectorize = vectorize, na.rm = na.rm)
> expected <- lubridate::dhours(30)
> expect_equal(object, expected) # 30:00:00
> ```
>
> And then most of those variables get changed later (some do stay the same, of course). I find it much easier to follow tests like these if you do repeat the values and put them into the call:
>
> ```
> object <- sum_time(t, u, v, w, x, y, z, class = "duration", circular = FALSE,
>                    vectorize = FALSE, na.rm = FALSE)
> expected <- lubridate::dhours(30)
> expect_equal(object, expected) # 30:00:00
> ```
>
> or even:
>
> ```
> expect_equal(
>   sum_time(t, u, v, w, x, y, z, class = "duration", circular = FALSE,
>            vectorize = FALSE, na.rm = FALSE),
>   lubridate::dhours(30)
> ) # 30:00:00
> ```
>
> There's a bit more repetition of things like `FALSE` or `"duration"` but it's _much_ easier to read as someone trying to understand the code + it makes moving the tests around in this file (or to another file) muuuch easier because it's not reliant on what came before it.

__Response__: Agreed. Changes implemented in gipso/mctq@f0f9aaa.

> ##### The functionality of `midday_change()`
>
> This might be my own out-of-area expertise, but `midday_change()` seems like an odd function. After reading the docs, I expected it would give me the day I'm on + 1 day and not just after the epoch. Is this a term-of-art (or transform-of-art?) that I'm not familiar with in chronobiology?

__Response__: `midday_change()` is mostly used to represent time values in charts using a circular time frame of 24 hours. I changed the documentation of `qplot_walk()` to make it more clear (changes implemented in gipso/mctq@75a2d17).

This is better understood with an example.

Say you have a vector of time values that cross the midnight hour (e.g., a `hms` vector with `22:00`, `23:00`, `00:00`, `01:00` values) and you want to plot a histogram of it. Remember, time values are not attached to a timeline.

To make it easier, I will use a vector with the same characteristics mentioned above that can be found in the `std_mctq` dataset included in the `mctq` package.

``` r
library(mctq)

data <- mctq::std_mctq$bt_w

head(data, 10)
#> 22:30:00
#> 22:50:00
#> 21:25:00
#> 00:30:00
#> 22:25:00
#> 21:00:00
#>       NA
#> 22:30:00
#> 22:00:00
#> 23:00:00
```

If you plot a histogram with this time vector, it would look like this:

``` r
library(ggplot2)

ggplot2::qplot(data, xlab = "Time", ylab = "Frequency", bins = 30)
```

![Without using `midday_change ()`](https://user-images.githubusercontent.com/29462472/136733768-ef1732c2-1284-4adc-aefe-0f22e2fa1070.png)

Note that the chart above uses a linear time frame, creating a gap between the data.

In some cases, we need to fit the data into a circular time frame of 24h using those charts/representations. That's what `midday_change()` is for.

This function attaches the time values to a two-day timeline by using the midday hour as a cutting point. If the time value has 12h or more, `midday_change()` will place it on day 1. Else (if the time value has less than 12h), on day 2.

``` r
dplyr::tibble(before = head(data, 10),
              after = head(mctq:::midday_change(data), 10))
#> # A tibble: 10 x 2
#>    before after
#>    <time> <dttm>
#>  1 22:30  1970-01-01 22:30:00
#>  2 22:50  1970-01-01 22:50:00
#>  3 21:25  1970-01-01 21:25:00
#>  4 00:30  1970-01-02 00:30:00
#>  5 22:25  1970-01-01 22:25:00
#>  6 21:00  1970-01-01 21:00:00
#>  7    NA  NA
#>  8 22:30  1970-01-01 22:30:00
#>  9 22:00  1970-01-01 22:00:00
#> 10 23:00  1970-01-01 23:00:00
```

Now, if instead of using `data` we plot a histogram using `mctq:::midday_change(data)`, that's what we would get:

``` r
ggplot2::qplot(mctq:::midday_change(data), xlab = "Time", ylab = "Frequency", bins = 30)
```

![Using `midday_change ()`](https://user-images.githubusercontent.com/29462472/136733876-220aa172-f27e-4930-8fee-d498533897dd.png)

We can also use `midday_change()` to compute circular statistics (e.g., mean, sd), but it's usually not the best way to do this (see the [circular](https://cran.r-project.org/web/packages/circular/index.html) package to learn more about circular statistics).

You can think of `midday_change()` as a "cheap" way to find the shortest interval between more than 2 time values (for only 2 time values we can use  `mctq::shorter_interval()`). It's also important to remember that a model with a good fit doesn't necessarily mean that it's a good representation of the phenomenon studied. In the same way, the shortest interval may not be the "right" interval. Knowing or estimating the "right" interval would require a better look at the data and the theory behind it (similar to the [discussion of the social jetlag computation](https://gipso.github.io/mctq/articles/sjl-computation.html)). Since `mctq` deals with sleep data, using only the midday as a cutting point usually works well.

> ##### Assert the message (or a part of the message) on errors
>
> Suggestion: There are a number of `expect_error(function(arg, arg))` I've found it helpful + caught a number of bugs myself when I assert error message itself (or some relevant portion of it)

__Response__: Agreed. I implemented the match for the `checkmate` error messages (Changes implemented in gipso/mctq@f0f9aaa). Recently, I decided to use the [`cli`](https://cli.r-lib.org/) package to facilitate the creation of the command line interface, which creates some difficulties when trying to match the messages. That's because `cli` uses a type of formatting difficult to match, hence, some of the expected errors don't have a pattern.

> ##### test-utils.R
>
> lines 246-247 look like they are duplicated from the test above (and not relevant to `shush()`)

__Response__: My mistake. Changes implemented in gipso/mctq@fbaf698.

> ##### Simplifying `swap()`
>
> `swap()` could be simplified to `swap <- function(x, y) list(x = y, y = x)`, but if you don't want to I would recommend making `a` and `b` more explanatory (i.e. `first_arg` and `second_arg` respectively)
>
> `swap_if()` you should avoid parse/evaling since that can lead to some very very strange errors it looks like the one use + the tests, this is always a relation between the two arguments that are given. Why not make the condition either be a logical (that's evaluated naturally) like `swap_if(x, y, condition = x > y)` or, you could make the condition a test, something like `` swap_if(x, y condition = `>`) `` which inside of it calls `condition(x, y)` so that you can change the relation between the two arguments.

__Response__: You're right (🤦). Changes implemented in gipso/mctq@ac55804.

> ##### `clock_roll()` method?
>
> `clock_roll()` have you considered making this a method that dispatches based on the class of `x`?

__Response__: Agreed. Changes implemented in gipso/mctq@ae4f3a9.

> ##### `na_as()` have you considered makign this a method as well?

__Response__: Agreed. Changes implemented in gipso/mctq@3c563ef and gipso/mctq@367cef0.

> ##### `get_class()`
>
> I don't think that `|| is.data.frame(x)` is ever called since data.frames inherit from lists (e.g. `> is.list(data.frame(x=1, y=2)); [1] TRUE`). `foo` could use a slightly more descriptive name, and is there any reason you don't use it in the `else` branch?

__Response__: You're right. I removed the `is.data.frame(x)` call. I also removed the `foo` function, writing it inside the `vapply()` call. The reason that `foo` is not used in the `else` branch is that you don't need to use a functional approach when assessing the class of an atomic vector. That's not the case with recursive vectors (like `list`) since they  can have elements of different classes. Changes implemented in gipso/mctq@c404b07.

> ##### `str_extract_()`
>
> is there any reason not to use `...` to pass most of the arguments to `regexpr`?

__Response__: `str_extract_()` is a workaround to avoid dependencies. Is nothing more than a `stringr::str_extract()` or `stringi::stri_extract()` like function. One of the benefits of a `stringr::str_extract()` like function is that [the first argument is always the vector of strings to modify, which makes `stringr` work particularly well in conjunction with the pipe](https://stringr.tidyverse.org/#compared-to-base-r). Using an ellipsis to pass the arguments to `regexpr` would defeat this purpose.

> ##### `dialog_line()`
>
[…]
