### 2021-10-15 #434 @jooolia
very cool @danielvartan about the discussions spurred by your package and presentations. :)
