### 2021-10-17 #434 @danielvartan
Hi @jonkeane ,

About the `gu()` comment. Unfortunately, that same check is used in several other functions (e.g., `mctq::so()`, `mctq::se`). A change only for `gu()` wouldn't resolve this issue.

The `checkmate` error messages can be a little unfriendly, but at least it have a regular pattern. Changing one check breaks this pattern, and, I think, this can be more detrimental for the user. That's a trade-off for not having to built a lot of [error constructors](https://style.tidyverse.org/error-messages.html).
