### 2021-10-30 #434 @danielvartan
Thank you @jooolia, @jonkeane and @leocadio-miguel for all your attention and dedication in reviewing our package.

I learned a lot during this process and was very happy to be a part of and contribute to the rOpenSci initiative! I already have a few other package projects underway and will soon be submitting them to the same review process.

I've already transferred the package to rOpenSci and I'm now taking care of the final details. I will soon post the checklist here for you to follow this process.
