### 2021-10-30 #434 @danielvartan
All set. Thanks again!

---

To-dos:
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
- [x] We're starting to roll out software metadata files to all rOpenSci packages via the Codemeta initiative, see https://docs.ropensci.org/codemetar/for how to include it in your package, after installing the package - should be easy as running `codemetar::write_codemeta()` in the root of your package.
